Makes ssize_t windows typedef more appropriate. It should be defined as SSIZE_T.

+++ lib/libwebsockets.h
@@ -35,6 +35,7 @@ extern "C" {
 #include <winsock2.h>
 #include <ws2tcpip.h>
 #include <stddef.h>
+#include <BaseTsd.h>
 #include "websock-w32.h"
 
 #include "gettimeofday.h"
@@ -42,7 +43,7 @@ extern "C" {
 #define strcasecmp stricmp
 #define getdtablesize() 30000
 
-typedef int ssize_t;
+typedef SSIZE_T ssize_t;
 
 #define LWS_VISIBLE
